% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{build_graph}
\alias{build_graph}
\title{Build the Graph}
\usage{
build_graph(g, k)
}
\arguments{
\item{g}{an igraph network}

\item{k}{the minimum degree}
}
\value{
the filtered network
}
\description{
Build the Graph
}
