% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{extract_giant_component}
\alias{extract_giant_component}
\title{Extract the Giant Component}
\usage{
extract_giant_component(g)
}
\arguments{
\item{an}{igraph network object}
}
\value{
an igraph network representing the largest connected component
}
\description{
Extract the Giant Component
}
