% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_tools.R
\name{get_coef_table}
\alias{get_coef_table}
\title{Get a Table of Coefficient Estimates}
\usage{
get_coef_table(hergm_results, format_output = "default")
}
\arguments{
\item{hergm_results}{An estimation result returned by `lighthergm::hergm()`
with `estimate_coefficients = TRUE`.}

\item{format_output}{filename or object type (character string).
See the help page of `modelsummary::modelsummary()`.}
}
\value{
a `modelsummary_list` table containing the coefficient estimates
}
\description{
Get a Table of Coefficient Estimates
}
