% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_tools.R
\name{get_descriptive_stats}
\alias{get_descriptive_stats}
\title{Get Data Descriptive Statistics}
\usage{
get_descriptive_stats(nodes)
}
\arguments{
\item{nodes}{a dataframe containing node covariates}
}
\value{
a list containing industry, occupation and city counts
}
\description{
Get Data Descriptive Statistics
}
