% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{get_graph_stats}
\alias{get_graph_stats}
\title{Descriptive Statistics for the Graph}
\usage{
get_graph_stats(g)
}
\arguments{
\item{g}{an igraph network}
}
\value{
a list containing summary statistics of the network, the degree distribution and its summary
}
\description{
Descriptive Statistics for the Graph
}
