% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_tools.R
\name{get_nodes_with_block}
\alias{get_nodes_with_block}
\title{Attach Block Affiliations}
\usage{
get_nodes_with_block(nodes, blocks, g)
}
\arguments{
\item{nodes}{a dataframe containing node covariates}

\item{blocks}{a factor containing the block IDs for each node (in the same order as the one used in the network object)}

\item{g}{a statnet network object}
}
\value{
a tibble containing the node covariates and the block affiliations
}
\description{
Attach Block Affiliations
}
