% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_tools.R
\name{get_random_seed}
\alias{get_random_seed}
\title{Get Random Seed
Obtains a random seed from the environment or from the System time}
\usage{
get_random_seed()
}
\value{
an integer seed
}
\description{
Get Random Seed
Obtains a random seed from the environment or from the System time
}
