% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{keep_k_core}
\alias{keep_k_core}
\title{Keep k-core}
\usage{
keep_k_core(g, k)
}
\arguments{
\item{g}{an igraph network object}

\item{k}{the minimum degree}
}
\value{
an igraph network object containing nodes that belong to the k-core
}
\description{
Keep k-core
}
