% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{keep_tokyo}
\alias{keep_tokyo}
\title{Keep only Tokyo nodes}
\usage{
keep_tokyo(g)
}
\arguments{
\item{g}{An igraph network object}
}
\value{
the filtered network that includes only nodes located in Tokyo
}
\description{
Keep only Tokyo nodes
}
