% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{make_base_igraph_network}
\alias{make_base_igraph_network}
\title{Create the igraph object}
\usage{
make_base_igraph_network(graph_data, features)
}
\arguments{
\item{graph_data}{a list containing the edgelist and node attriutes}

\item{features}{a list of features to include in the graph}
}
\value{
an igraph network
}
\description{
Create a graph from node covariates and edgelists
}
