% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_tools.R
\name{plot_block_size}
\alias{plot_block_size}
\title{Plot the Block Size Distribution}
\usage{
plot_block_size(partition)
}
\arguments{
\item{partition}{a factor containing the ID of the block assigned to each node}
}
\value{
a ggplot2 plot of the distribution of block sizes
}
\description{
Plot the Block Size Distribution
}
