% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_tools.R
\name{plot_lower_bound}
\alias{plot_lower_bound}
\title{Plot the Lower Bound}
\usage{
plot_lower_bound(lower_bound)
}
\arguments{
\item{lower_bound}{a lower bound vector returned by `lighthergm_res$EM_lower_bound`}
}
\value{
a ggplot2 plot of the evolution of the lower bound with respect to the iteration number
}
\description{
Plot the Lower Bound
}
