% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_tools.R
\name{save_plot}
\alias{save_plot}
\title{Save Plot}
\usage{
save_plot(plot, file_path)
}
\arguments{
\item{plot}{a ggplot2 plot to save}

\item{file_path}{the path where to store the plot}
}
\description{
Save Plot
}
